+++ src/Symfony/Component/HttpKernel/DataCollector/DataCollectorInterface.php
@@ -18,6 +18,8 @@ use Symfony\Component\HttpFoundation\Response;
  * DataCollectorInterface.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @method reset() Resets this data collector to its initial state.
  */
 interface DataCollectorInterface
 {