+++ src/Symfony/Component/HttpKernel/Tests/Fixtures/TestEventDispatcher.php
@@ -25,4 +25,8 @@ class TestEventDispatcher extends EventDispatcher implements TraceableEventDispa
     {
         return array('bar');
     }
+
+    public function reset()
+    {
+    }
 }